containerfiles: dcap-registration-flow: add missing enclave packages
The contents of sgx-pck-id-retrieval-tool-1.23.100.0-1.el9.x86_64 RPM
suggests that it has the necessary enclaves installed:

/opt/intel/sgx-pck-id-retrieval-tool
/opt/intel/sgx-pck-id-retrieval-tool/License.txt
/opt/intel/sgx-pck-id-retrieval-tool/PCKIDRetrievalTool
/opt/intel/sgx-pck-id-retrieval-tool/README.txt
/opt/intel/sgx-pck-id-retrieval-tool/libsgx_id_enclave.signed.so.1
/opt/intel/sgx-pck-id-retrieval-tool/libsgx_pce.signed.so.1
/opt/intel/sgx-pck-id-retrieval-tool/network_setting.conf
/usr/lib/.build-id
/usr/lib/.build-id/a4
/usr/lib/.build-id/a4/29e221420f6830e9fe461c90819c2f50e8e980

but that does not seem to be enough for the tool to function. Running
the tool errors with:

Error, call sgx_create_enclave: fail [load_enclave], SGXError:200f.

It does work if the two enclaves are installed from their own RPM
packages: libsgx-ae-id-enclave-1.23.100.0-1.el9.x86_64 and
libsgx-ae-pce-2.26.100.0-1.el9.x86_64.

Signed-off-by: Mikko Ylinen <mikko.ylinen@intel.com>