DCAP: don't use original tarball ownership
The UIDs/GIDs in the tarball come from a different system and are very
large (like the 11525723 in your error), which are not available or
are outside the allowed range in the build container's user namespace.

To fix this we can use '--no-same-owner' as an argument for the tar
command.

Signed-off-by: Daniel Kreling <dkreling@redhat.com>