osc-pcss: do not use "latest" for the builder and base image tags
Using "latest" in our builds can lead to silent changes and breaks.
We should always use a specific tag for the images we use, knowing that
we have a Mintmaker process that will update them on a regular basis,
making a pull request that we can review and validate before merging.

Signed-off-by: Julien Ropé <jrope@redhat.com>