osc-pccs: build sqlite3 bindings from source using nodejs-devel headers
sqlite3 npm package uses native bindings which are prebuilt
and made available as part of the release.

npm install uses the "prebuild" tool to get the artifacts downloaded
but this fails in the hermetic build environment (hermeto prefetch for
npm does not seem to be able to handle this case).

While a prefetch would be possible, this violates enterprise contract
sanity checks for the build artifacts.

Instead, build the bindings using the nodejs headers provided
by nodejs:22/development. With this, we avoid prefetching any
external dependencies.

Signed-off-by: Mikko Ylinen <mikko.ylinen@intel.com>