osc-pccs: use UBI10.1 as builder/base image
This gives us access to the RHEL repositories where some of the packages we
need are located.

Also stop using "latest" in our build, as it can lead to silent changes and
breaks. We should always use a specific tag for the images we use, knowing that
we have a Mintmaker process that will update them on a regular basis, making a
pull request that we can review and validate before merging.

Signed-off-by: Julien Ropé <jrope@redhat.com>