konflux: update the rpms-signature-scan image reference
We have Enterprise Contracts failures for pccs and tdx image because they are
using an older version of this task.
The automated update from konflux is not updating this task, so this commit is
doing manually based on the SHA provided by the error message in the Enterprise
Contract.

Also modifying the storage-helper pipeline, even if it is not impacted (was using
a newer image reference apparently).

Signed-off-by: Julien Ropé <jrope@redhat.com>